Support reading and writing STRING
Strings are handled as a CIP_STRUCT with a certain STRING identifier.
Unfortunately the 'elements' (always 1) is embedded when writing but not
reading, complicating the code for handling the CIPData of type
STRUCT.STRING.

Also added 'nls' settings, and encode() now throws an exception because
of potential issues with string type.

#12