Add LICENSE file
On GitHub, the convention is to include a LICENSE file in
your project, which holds the license under which the code
can be used. Since the header comments in the source files
specify that they are released under the Eclipse Public
License v1.0, I added that license in a new LICENSE file.

The text for the LICENSE file was copied from
https://choosealicense.com/licenses/epl-1.0/.